+++ local-s3-core/src/main/java/com/robothy/s3/core/service/BucketService.java
@@ -13,7 +13,8 @@ import java.util.stream.Collectors;
 
 public interface BucketService extends CreateBucketService, BucketVersioningService, BucketTaggingService,
     BucketAclService, BucketPolicyService, BucketReplicationService,
-    BucketEncryptionService, BucketPublicAccessBlockService, BucketPolicyStatusService, BucketCorsService {
+    BucketEncryptionService, BucketPublicAccessBlockService, BucketPolicyStatusService, BucketCorsService,
+    BucketLifecycleService {
 
   /**
    * Delete a bucket.